requirements.txt
PyMuPDF==1.23.21
Pillow==10.2.0
requests==2.31.0
